+++ data/data-authentication/src/main/java/com/bluehabit/eureka/data/authentication/datasource/remote/request/SetNewPasswordRequest.kt
@@ -1,0 +1,12 @@
+/*
+ * Copyright © 2023 Blue Habit.
+ *
+ * Unauthorized copying, publishing of this file, via any medium is strictly prohibited
+ * Proprietary and confidential
+ */
+
+package com.bluehabit.eureka.data.authentication.datasource.remote.request
+
+data class SetNewPasswordRequest (
+    val password: String,
+)
