+++ data/data-authentication/src/main/java/com/bluehabit/eureka/data/authentication/datasource/remote/response/SetNewPasswordResponse.kt
@@ -5,8 +5,5 @@
  * Proprietary and confidential
  */
 
-package com.bluehabit.eureka.data.authentication.datasource.remote.request
+package com.bluehabit.eureka.data.authentication.datasource.remote.response
 
-data class SetNewPassword (
-    val password: String,
-)
